+++ spec/actions/readSpreadsheet.spec.js
@@ -1,15 +1,18 @@
+/* eslint-disable no-unused-expressions */
 const fs = require('fs');
 const chai = require('chai');
 const chaiAsPromised = require('chai-as-promised');
 const nock = require('nock');
 const sinon = require('sinon');
-const logger = require('@elastic.io/component-logger')();
-const createSpreadsheet = require('../../lib/actions/createSpreadsheet');
+
+const log = require('@elastic.io/component-logger')();
+
+const createSpreadsheetRow = require('../../lib/actions/createSpreadsheetRow');
 
 chai.use(chaiAsPromised);
 const { expect } = chai;
 
-xdescribe('Add new spreadsheet', () => {
+xdescribe('Read spreadsheet', () => {
   let emitter;
 
   let configuration;
@@ -30,67 +33,132 @@ xdescribe('Add new spreadsheet', () => {
       },
     },
   };
-
-  const wrongSecret = {
-    data: {
-      attributes: {
-        credentials: {
-          access_token: 'wrong',
-          refresh_token: 'refresh_token',
-          expires_in: 3599,
-          scope: 'https://www.googleapis.com/auth/spreadsheets,https://www.googleapis.com/auth/drive.metadata.readonly',
-          additional_params: '{"access_type":"offline","prompt":"consent"}',
-        },
-      },
-    },
-  };
   const secretId = 'secretId';
+  afterEach(() => {
+    nock.cleanAll();
+  });
   before(() => {
     if (fs.existsSync('.env')) {
       // eslint-disable-next-line global-require
       require('dotenv').config();
     }
     configuration = {
+      spreadsheetId: 'some_id',
+      worksheetId: 'some_worksheet',
+      mode: 'array',
       secretId,
-      retries: 5,
     };
   });
 
   beforeEach(() => {
     emitter = {
       emit: sinon.spy(),
-      logger,
+      logger: log,
     };
     nock(process.env.ELASTICIO_API_URI)
-      .get(`/v2/workspaces/${process.env.ELASTICIO_WORKSPACE_ID}/secrets/${secretId}`)
-      .reply(200, wrongSecret)
       .get(`/v2/workspaces/${process.env.ELASTICIO_WORKSPACE_ID}/secrets/${secretId}`)
       .reply(200, secret);
   });
 
-  afterEach(() => {
-    nock.cleanAll();
-  });
-
   it('success', async () => {
-    nock('https://sheets.googleapis.com')
-      .post('/v4/spreadsheets')
-      .reply(401, { status: 'OK' })
-      .post('/v4/spreadsheets')
-      .reply(200, { status: 'OK' });
-
     const msg = {
       body: {
-        properties: { title: 'Some name' },
-        sheets: [
-          { properties: { title: 'Sheet A' } },
-          { properties: { title: 'Sheet B' } },
-        ],
+        values: [1, -6.8, 'string_line', true],
       },
     };
 
-    const result = await createSpreadsheet.process.call(emitter, msg, configuration);
+    nock('https://sheets.googleapis.com')
+      .persist()
+      .post(
+        `/v4/spreadsheets/${configuration.spreadsheetId}/values/${configuration.worksheetId}:append?valueInputOption=RAW`,
+        { majorDimension: 'ROWS', values: [[1, -6.8, 'string_line', true]] },
+      )
+      .reply(200, { success: 'OK' });
 
-    expect(result.body).to.deep.equal({ status: 'OK' });
+    const result = await createSpreadsheetRow.process.call(emitter, msg, configuration);
+    expect(result.body).to.deep.equal({ success: 'OK' });
+  });
+
+  it('list Worksheets', async () => {
+    nock('https://sheets.googleapis.com')
+      .get(`/v4/spreadsheets/${configuration.spreadsheetId}`)
+      .reply(200, {
+        sheets: [
+          {
+            properties: {
+              sheetId: 1,
+              title: 'Sheet1',
+            },
+          },
+          {
+            properties: {
+              sheetId: 2,
+              title: 'Sheet2',
+            },
+          },
+        ],
+      });
+
+    const result = await createSpreadsheetRow.listWorksheets.call(emitter, configuration);
+    expect(result).to.deep.equal({ Sheet1: 'Sheet1', Sheet2: 'Sheet2' });
+  });
+
+  it('Generates metadata for array mode', async () => {
+    configuration.mode = 'array';
+    const result = await createSpreadsheetRow.getMetaModel.call(emitter, configuration);
+    expect(result.in).to.exist;
+    expect(result.out).to.exist;
+    expect(result.in.properties.values).to.be.exist;
+  });
+  it('Generates metadata for first header mode', async () => {
+    configuration.mode = 'header';
+    nock('https://sheets.googleapis.com').get(`/v4/spreadsheets/${configuration.spreadsheetId}/values/${configuration.worksheetId}`)
+      .reply(200, {
+        values: [['header1', 'header 2'], ['value1', 'value2']],
+      });
+    const result = await createSpreadsheetRow.getMetaModel.call(emitter, configuration);
+    expect(result.in).to.exist;
+    expect(result.out).to.exist;
+    expect(result.in.properties.header1).to.be.exist;
+    expect(result.in.properties.header2).to.be.exist;
+  });
+  it('Header mode error if values not present in first row', async () => {
+    configuration.mode = 'header';
+    nock('https://sheets.googleapis.com').get(`/v4/spreadsheets/${configuration.spreadsheetId}/values/${configuration.worksheetId}`)
+      .reply(200, {
+        values: [[], ['value1', 'value2']],
+      });
+    try {
+      await createSpreadsheetRow.getMetaModel.call(emitter, configuration);
+      expect(true, 'should fail').to.be.false;
+    } catch (e) {
+      expect(e.message).to.be.eql('Input Mode: "First Row As Headers" requires first row to have at least one cell with value. Check: Common Errors section in docs.');
+    }
+  });
+  it('Header mode throw error if values not unique', async () => {
+    configuration.mode = 'header';
+    nock('https://sheets.googleapis.com').get(`/v4/spreadsheets/${configuration.spreadsheetId}/values/${configuration.worksheetId}`)
+      .reply(200, {
+        values: [['header1', 'header1'], ['value1', 'value2']],
+      });
+    try {
+      await createSpreadsheetRow.getMetaModel.call(emitter, configuration);
+      expect(true, 'should fail').to.be.false;
+    } catch (e) {
+      expect(e.message).to.be.eql('Input Mode: "First Row As Headers" requires cells in first row to be unique. Check: Common Errors section in docs.');
+    }
+  });
+  it('Header mode throw errors if values is blank in first mode', async () => {
+    configuration.mode = 'header';
+    nock('https://sheets.googleapis.com').get(`/v4/spreadsheets/${configuration.spreadsheetId}/values/${configuration.worksheetId}`)
+      .reply(200, {
+        values: [['header1', '', 'header3'], ['value1', 'value2']],
+      });
+    try {
+      await createSpreadsheetRow.getMetaModel.call(emitter, configuration);
+      expect(true, 'should fail').to.be.false;
+    } catch (e) {
+      expect(e.message).to.be.eql('Input Mode: "First Row As Headers" requires cells in first row to be not empty. Check: Common Errors section in docs.');
+    }
   });
 });