+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 3.0.0 (June 20, 2022)
 
 * Use Faceless service for authentication 
+* Added `Read Spreadsheet` action
  
 ## 2.0.3 (May 20, 2022)